Consolidate pipeline utilities into single module
